Move duplicate volume merging functionality to separate file
Move logic for merging into a separate file and add a more abstract
interface to allow adding additional merging later if necessary.
